Fix usage of fast-glob API
globby depends on fast-glob which introduced a breaking change in 3.3.0. Their API has been expecting an string[] for `ignore` option for awhile but in the latest version it causes a TypeError.
Make sure the `ignore` option is passed an array.

Fixes issues for: https://github.com/bahmutov/cypress-split/issues/78#issuecomment-1630108191